feat: #89 coredata 모델 만들기
